third_party/blink/web_tests/platform/generic/fast/events/click-range-slider-expected.txt

Test that click events are fired for a slider when the range of values is dense enough that the thumb element is repositioned under the cursor on mouse press.

PASS successfullyParsed is true

TEST COMPLETE

PASS clickCount is 3
